Research_AI_law/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "research-ai-law"
version = "0.1.0"
description = "Estudio comparativo cross-country sobre regulación de IA y desarrollo del ecosistema IA"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "pandera>=0.18",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "rapidfuzz>=3.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
